requirements.txt
packaging==24.2
numpy==1.23
opencv-python==4.11.0.86
matplotlib==3.10.0
einops==0.6.1
setuptools==59.5.0
timm==1.0.15
PyWavelets==1.8.0
pytorch-wavelets==1.3.0
tensorboard==2.19.0